3rd Commit
So I didn't change too much here language-wise, but it was a huge jump from a module script (which have the same properties as a server script so I'm just going to refer to it as such) to a local script. This had to be done because of the other file committed, PlaybackLoudnessDetector. This property was giving me huge issues as it can only be accessed locally and then communicated to the server from there. For some reason, I couldn't get it to consistently communicate with the server. I decided to change MainScript to a local script bc then I could access the playbackloudness property without having to worry about first detecting it then sending it to the server. This makes my code much more "messy" but in many ways more efficient.